fix: add Offset, count to SearchShort function; change tests accordingly to SearchShort function; change int types in Page struct according to db.count function.

+++ internal/dto/page.go
@@ -5,5 +5,5 @@ type Page[T any] struct {
 	Data    []T
 	Page    int
 	PerPage int
-	Total   int
+	Total   int64
 }
